JSON serialization refactor (#4331)
