+++ src/mesh/http/ContentHandler.cpp
@@ -9,8 +9,8 @@
 #if HAS_WIFI
 #include "mesh/wifi/WiFiAPClient.h"
 #endif
-#include "mqtt/JSON.h"
 #include "power.h"
+#include "serialization/JSON.h"
 #include "sleep.h"
 #include <FSCommon.h>
 #include <HTTPBodyParser.hpp>
