+++ src/serialization/MeshPacketSerializer.h
@@ -1,0 +1,8 @@
+#include <meshtastic/mesh.pb.h>
+#include <string>
+
+class MeshPacketSerializer
+{
+  public:
+    static std::string JsonSerialize(meshtastic_MeshPacket *mp, bool shouldLog = true);
+};
